refactor: scope legacy job responses to API v3
