+++ flexmeasures/api/v3_0/sensors.py
@@ -41,10 +41,8 @@ from flexmeasures.api.common.schemas.sensor_data import (  # noqa F401
 from flexmeasures.api.common.schemas.sensors import SensorId  # noqa F401
 from flexmeasures.api.common.schemas.users import AccountIdField
 from flexmeasures.api.common.rate_limiting import limit_triggers
-from flexmeasures.api.common.utils.api_utils import (
-    process_sensor_data_ingestion,
-    use_legacy_job_responses,
-)
+from flexmeasures.api.common.utils.api_utils import process_sensor_data_ingestion
+from flexmeasures.api.v3_0.utils import use_legacy_job_responses
 from flexmeasures.data.services.utils import job_status_description
 from flexmeasures.api.common.utils.deprecation_utils import (
     _add_headers as add_deprecation_header,
@@ -679,9 +677,9 @@ class SensorAPI(FlaskView):
         response, code = process_sensor_data_ingestion(
             sensor_id=sensor.id,
             user_id=current_user.id,
-            asset=sensor.generic_asset,
             uploaded_files=files_for_job,
             upload_data=upload_data,
+            force_synchronous=use_legacy_job_responses(sensor.generic_asset),
         )
         return response, code
 
@@ -755,8 +753,8 @@ class SensorAPI(FlaskView):
         response, code = process_sensor_data_ingestion(
             sensor_id=sensor.id,
             user_id=current_user.id,
-            asset=sensor.generic_asset,
             sensor_data=sensor_data,
+            force_synchronous=use_legacy_job_responses(sensor.generic_asset),
         )
         return response, code
